finally get auto testdata detection working.
Can't run IO in Spec, so instead we allow TH to scan testdata
directory and construct Spec tree from there.

It's a bit unfortunate though as haskell-language-server seems to
always crash for any source code in test/, but as long as it doesn't
crash inside src/, we can live with it.