cut some tests.
let's not mix role of testing and benchmarking,
for now we seem to have a correct implementation,
and I'd say login p1 is short but sufficient
for verifying correctness.

moreover, I don't think there are much room of improvement left,
unless we can make some fundamental changes to this algorithm.